+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.0
+
+- Update dependency `typescript` to `3.7.2`, include support for TypeScript 3.7 features.
+- Add type `Awaited` to retrieve the realised promise type.
+- Remove `docs` from npm package
+- Add `unwrapJson` to `UnitConversion`.
+- Add string comparison.
+- Add `Option.str` for template string mapping.
+
 ## 1.1.9
 
 - `MemoryCache` uses `lru-cache` instead.